Challenge_V/alaska_snow_agent/requirements.txt
# Web Framework
streamlit==1.35.0

# Google Cloud Services
google-cloud-bigquery==3.17.2
google-cloud-aiplatform>=1.52.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0

# Data Processing
pandas>=2.0.0
db-dtypes>=1.0.0
numpy>=1.21.0

# Testing
pytest>=7.0.0
unittest-xml-reporting>=3.2.0

# NLP and Evaluation
nltk>=3.8.0
transformers>=4.21.0
rouge-score>=0.1.2
torch>=1.12.0
evaluate>=0.4.0
bert-score>=0.3.13
sentence-transformers>=2.2.0

# Utilities
tabulate>=0.9.0
python-dotenv>=0.19.0
pydantic>=1.10.0